+++ system/manager/process_config.py
@@ -77,7 +77,7 @@ procs = [
   PythonProcess("torqued", "selfdrive.locationd.torqued", only_onroad),
   PythonProcess("controlsd", "selfdrive.controls.controlsd", not_joystick),
   PythonProcess("joystickd", "tools.joystick.joystickd", joystick),
-  PythonProcess("longitudinal_profilesd", "examples.longitudinal_profiles", long_maneuver),
+  PythonProcess("maneuversd", "examples.maneuversd", long_maneuver),
   PythonProcess("selfdrived", "selfdrive.selfdrived.selfdrived", only_onroad),
   PythonProcess("card", "selfdrive.car.card", only_onroad),
   PythonProcess("deleter", "system.loggerd.deleter", always_run),
